+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Added
 * Added `EthernetClient::localIP()`.
 * Added `EthernetClass::hostByName(hostname, ip)` convenience function.
+* Added `EthernetClass::setDNSServerIP(index, ip)` and `dnsServerIP(index)`.
 
 ### Changed
 * Updated lwIP to the latest master (5e3268cf).